toolbox/modprobe: update kernel version directory filter
For android15-6.6 and later kernels, for 4kb page size kernel variants
include the substring "-4k". The 16kb page size kernels omit this
substring. So update modprobe to check for the `-4k` substring when the
kernel modules are under the folder `/lib/modules/$UTS_RELEASE` for
android15-6.6 and later.

This fixes loading pkvm modules when booting kernel images packaged from
the kernel tree. For Android platform images, the kernel modules are
directly under `/lib/modules` and therefore _pgagnostic targets are not
affected by this.

Test: boot sb3 with v6.1
Test: boot sb3 with v6.12 and 16kb pages
Bug: 340631213
Change-Id: I657ec9cf5cefd7d9831a7f5d2453fb21acc6d3be